feat(functional minimize button): adding a minimize button to any modal now makes it functional
If you add <TitleBar.Minimize key="minimize" /> to the titlebar of any modal it will now minimize it
to the TaskBar. It's vital that you do have a TaskBar so that you can restore the modal, if you want
to use this feature. You also HAVE to add the key="minimize" for it to work. There's no styling to
this feature yet, for now the modal is just hidden and restored, however it's something I might work
on in the future (unless someone else wants to!).